Add theme manager
Added theme manager that can change currently selected theme on the fly. It doesn't work as it
should but basic functionality is there. Later it will use directory name to determine theme,
which will allow custom icons, rescources etc. in that directory. Fixed memory errors when allocating
space for configuration and theme variables. Small changes to mvc and makefiles.