Fix memory leaks
Fixed multiple memory leaks, read/write errors and dangling pointers in theme and config. Added Manager
files and structure, which is used as a base for theme and config managers. Their functions are now
ready and should be used when creating managers. Only official functions should be used to avoid memory leaks
and maintain structural integrity. There is still work to be done with model, which doesn't yet correctly
allocate and free memory.