Add json parsing utilities
Added new json parser, which is capable to create json-object. This allows easier
way to manipulate values in an object. It also makes it easier to handle exceptions.
Json utilities are still unfinished, but provide better access than previous read
and write system. Model was adapted to use this new interface.