Cleanup signal handlers
Cleaned up some code in singal.c file. Changed function names in glade files to be more
unified. Fixed some errors in model and view and added small updates to ViewObject functions.
Added zoom value to read and write  functions. It is used to determine size of the cells on screen.