Add new grid update functions.
Added new structures for and functions for updating grid information. New model
only keeps track of cells that are alive and cells close to them. It is functioning
but uses a lot of rescources when a lot of cells are alive. However when finnished
this should be more efficient way than checking every cell each turn.